+++ pkg/appsec/security_policy_test.go
@@ -92,3 +92,87 @@ func TestAppSec_ListSecurityPolicies(t *testing.T) {
 		})
 	}
 }
+
+func TestAppSec_CreateSecurityPolicyWithDefaultProtectionsSuccess(t *testing.T) {
+
+	response := CreateSecurityPolicyResponse{}
+	respData := compactJSON(loadFixtureBytes("testdata/TestSecurityPolicy/CreateSecurityPolicyWithDefaultProtectionsResponse.json"))
+	err := json.Unmarshal([]byte(respData), &response)
+	require.NoError(t, err)
+
+	requestParams := CreateSecurityPolicyWithDefaultProtectionsRequest{
+		ConfigVersion: ConfigVersion{
+			ConfigID: 43253,
+			Version:  15,
+		},
+		PolicyName:   "akamaitools",
+		PolicyPrefix: "AK01",
+	}
+
+	tests := map[string]struct {
+		params           CreateSecurityPolicyWithDefaultProtectionsRequest
+		responseStatus   int
+		responseBody     string
+		expectedMethod   string
+		expectedPath     string
+		expectedResponse *CreateSecurityPolicyResponse
+		withError        error
+		headers          http.Header
+	}{
+		"200 OK": {
+			params: requestParams,
+			headers: http.Header{
+				"Content-Type": []string{"application/json"},
+			},
+			responseStatus:   http.StatusOK,
+			responseBody:     string(respData),
+			expectedMethod:   http.MethodPost,
+			expectedPath:     "/appsec/v1/configs/43253/versions/15/security-policies/protections",
+			expectedResponse: &response,
+		},
+		"500 internal server error": {
+			params:         requestParams,
+			headers:        http.Header{},
+			responseStatus: http.StatusInternalServerError,
+			responseBody: `
+{
+    "type": "internal_error",
+    "title": "Internal Server Error",
+    "detail": "Error creating security policy with default protections",
+    "status": 500
+}`,
+			expectedMethod: http.MethodPost,
+			expectedPath:   "/appsec/v1/configs/43253/versions/15/security-policies/protections",
+			withError: &Error{
+				Type:       "internal_error",
+				Title:      "Internal Server Error",
+				Detail:     "Error creating security policy with default protections",
+				StatusCode: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mockServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				assert.Equal(t, test.expectedPath, r.URL.String())
+				assert.Equal(t, test.expectedMethod, r.Method)
+				w.WriteHeader(test.responseStatus)
+				_, err := w.Write([]byte(test.responseBody))
+				assert.NoError(t, err)
+			}))
+			client := mockAPIClient(t, mockServer)
+			_, err := client.CreateSecurityPolicyWithDefaultProtections(
+				session.ContextWithOptions(
+					context.Background(),
+					session.WithContextHeaders(test.headers),
+				),
+				test.params)
+			if test.withError != nil {
+				assert.True(t, errors.Is(err, test.withError), "want: %s; got: %s", test.withError, err)
+				return
+			}
+			require.NoError(t, err)
+		})
+	}
+}